bot_project/requirements.txt
requests==2.31.0
pandas==2.2.3
statsmodels==0.14.4
python-slugify==8.0.4
python-telegram-bot==20.6
numpy==1.26.2
aiohttp==3.9.1
python-dotenv==1.0.1
httpx==0.25.0
